# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for getting song lyrics with translation by artist and title",
        "title": "Lyrics Library API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/register": {
            "post": {
                "description": "Register a new user with email and password",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "Registration data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "400": {
                        "description": "Invalid request data",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "User already exists",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/track": {
            "get": {
                "description": "If 'title' is provided, returns lyrics for the specific song.\nOtherwise, returns a list of all songs by the artist (without track).",
                "tags": [
                    "track"
                ],
                "summary": "Get song lyrics or artist tracks",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"Juice WRLD\"",
                        "description": "Artist name",
                        "name": "artist",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"Legends\"",
                        "description": "Song title (optional)",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns lyrics (object) or artist tracks (array)",
                        "schema": {
                            "$ref": "#/definitions/dto.TrackResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Save lyrics and translation for a given artist and song title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "track"
                ],
                "summary": "Save a new track with translation",
                "parameters": [
                    {
                        "description": "Lyrics request data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully saved track",
                        "schema": {
                            "$ref": "#/definitions/dto.TrackResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/track/{uuid}": {
            "delete": {
                "description": "Delete song lyrics by uuid",
                "tags": [
                    "track"
                ],
                "summary": "Delete song lyrics",
                "parameters": [
                    {
                        "type": "string",
                        "example": "e434dc13-ada5-4bde-b695-d97014dadebc",
                        "description": "Track UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Lyrics deleted successfully"
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateRequest": {
            "type": "object",
            "required": [
                "artist",
                "title"
            ],
            "properties": {
                "artist": {
                    "type": "string",
                    "example": "Juice WRLD"
                },
                "title": {
                    "type": "string",
                    "example": "Lucid Dreams"
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dto.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "test@test.com"
                },
                "password": {
                    "type": "string",
                    "example": "matveyisgoat123"
                }
            }
        },
        "dto.TrackResponse": {
            "type": "object",
            "properties": {
                "artist": {
                    "type": "string",
                    "example": "Lucid Dreams"
                },
                "lyrics": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "I still see your shadows in my room..."
                    ]
                },
                "title": {
                    "type": "string",
                    "example": "Juice WRLD"
                },
                "translation": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Я все еще вижу твои тени в моей комнате..."
                    ]
                }
            }
        }
    }
}